Tidy up the terminology in the index handler
