+++ src/hotchocolate/utils.py
@@ -1,5 +1,6 @@
 # -*- encoding: utf-8
 
+import collections
 import filecmp
 import os
 import re
@@ -41,3 +42,42 @@ def lazy_copyfile(src, dst):
             return
     os.makedirs(os.path.dirname(dst), exist_ok=True)
     shutil.copyfile(src, dst)
+
+
+Pageset = collections.namedtuple('Pageset', ['slug', 'posts', 'next', 'prev'])
+
+
+class Pagination:
+    """
+    Provides pagination for a series of posts.
+
+    Given a series of posts, the page size and the URL prefix
+    (i.e. what's the URL slug for this series of posts), you can iterate
+    over this object to get the collection of posts, slug and next/prev links
+    for these posts.
+    """
+    def __init__(self, posts, page_size, prefix):
+        self.posts = posts
+        self.size = page_size
+        self.prefix = prefix.strip('/')
+
+    def __iter__(self):
+        for pageno, posts in enumerate(chunks(self.posts, self.size), start=1):
+            if pageno == 1:
+                if self.prefix:
+                    slug = '/%s/' % self.prefix
+                else:
+                    slug = '/'
+                prev_ = None
+            else:
+                prefix = self.prefix or 'page'
+                prev_ = slug
+                slug = '/%s/%d/' % (prefix, pageno)
+
+            if self.size * pageno >= len(self.posts):
+                next_ = None
+            else:
+                prefix = self.prefix or 'page'
+                next_ = '/%s/%d/' % (prefix, pageno + 1)
+
+            yield Pageset(slug, posts, next_, prev_)
